refactor: Remove unmeaningful components
Remove unmeaningful components: Button.jsx, Title.jsx
