--- kakao-shopping/src/components/atoms/Title.jsx
@@ -1,4 +1,0 @@
-const Title = ({children, className}) => { // 내용 텍스트, 클래스명
-  return <h2 className={className}>{children}</h2>
-}
-export default Title;
