+++ kakao-shopping/src/components/molecules/SelectedOption.jsx
@@ -1,5 +1,4 @@
 import { useState } from "react";
-import Button from "../atoms/Button";
 import { useDispatch } from "react-redux";
 import { addItem, subtractItem } from "../../redux/cartRedux";
 
@@ -42,19 +41,19 @@ const SelectedOption = ({ selectedOption, className }) => {
         <div className="w-4/5 border-t mx-auto my-2"/>
         <div className="flex justify-between">
           <div className="flex my-2 items-center">
-            <Button 
+            <button 
               onClick={() => handleCountClick(-1)}
               className={`${buttonValid ? '' : 'text-gray-200'} border mx-1 bg-gray-100 border-gray-500 rounded-md w-5 h-5 flex justify-center items-center`}
             >
               <div>-</div>
-            </Button>
+            </button>
             <div className="h-full text-center w-5 mx-auto">{optionCount}</div>
-            <Button 
+            <button 
               onClick={() => handleCountClick(1)}
               className="border mx-1 bg-gray-100 border-gray-500 rounded-md w-5 h-5 flex justify-center items-center"
             >
               <div>+</div>
-            </Button>
+            </button>
           </div>
           <div className="my-2">
             {selectedOption.price}원
