+++ kakao-shopping/src/components/organisms/LoginForm.jsx
@@ -1,5 +1,4 @@
 import InputGroup from "../molecules/InputGroup"
-import Button from "../atoms/Button";
 import useInput from "../../hooks/useInput";
 import { loginApi } from "../../apis/api";
 import { useDispatch } from "react-redux";
@@ -70,17 +69,17 @@ const LoginForm = () => {
       <span className="text-red-500 text-xs block h-2">{loginFailed}</span>
       
       {everythingIsValid ? 
-      <Button
+      <button
         className="w-96 m-2 pr-6 pl-6 p-2 bg-yellow-300 hover:bg-yellow-400 rounded-md mb-10"
         onClick={handleLogin}>
           로그인
-      </Button> :
+      </button> :
         
-      <Button
+      <button
         className="w-96 m-2 pr-6 pl-6 p-2 bg-yellow-300 text-gray-400 cursor-default rounded-md mb-10"
         >
           로그인
-      </Button>}
+      </button>}
     <div className="absolute left-2 bottom-0">
       <Link to='/register'><span className="text-sm">회원가입</span></Link>
       </div>
