+++ kakao-shopping/src/components/organisms/ProductDetail.jsx
@@ -1,5 +1,4 @@
-import StarRating from "./StarRating"
-import Button from "../atoms/Button"
+import StarRating from "../atoms/StarRating"
 import ProducePrice from "../atoms/ProductPrice"
 
 const ProductDetail = ({product}) => {
@@ -9,13 +8,13 @@ const ProductDetail = ({product}) => {
       <div className="w-96">
         <StarRating starCount={product.starCount} />
         <div className="pb-1 text-xl">{product.productName}</div>
-        <Button className="bg-yellow-300 rounded-full flex flex-column align-center w-32 py-3 justify-center">
+        <button className="bg-yellow-300 rounded-full flex flex-column align-center w-32 py-3 justify-center">
           <div className="flex font-bold text-sm">
             <span>톡딜가</span>
             <ProducePrice price={product.price} className="mx-1" />
             <span> 원 ~</span>
           </div>
-        </Button>
+        </button>
       </div>
     </div>
   )
